introduce CliCoreBase.ts to remove all Node repl dependencies, allowing future REPL to be built without using Node REPL. Small change to stringify so that Timer objects are recognized in Bun. Add couple Bun detectors in tests to work around Jest specialties that are not availavle in Bun. Add github actions so that all tests run on Bun as well as node.
