+++ package.json
@@ -26,6 +26,7 @@
     "build": "tsc && npm run webpack && npm run docs",
     "webpack": "webpack && webpack --config webpack.config.cjs.js && webpack --config webpack.nodeSingleBin.cjs.js",
     "test": "npm run clean && npm run build && node --experimental-vm-modules node_modules/jest/bin/jest.js --detectOpenHandles",
+    "test:bun": "bun run test",
     "test-fast": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
     "stated": "node --experimental-vm-modules dist/stated.js",
     "docs": "typedoc --out docs src/*"
@@ -54,11 +55,11 @@
     "flatbuffers": "^23.5.26",
     "js-yaml": "^4.1.0",
     "json-colorizer": "^2.2.2",
-    "jsonata": "^2.0.3",
+    "jsonata": "2.0.4",
     "lodash": "^4.17.21",
     "lodash-es": "^4.17.21",
     "minimist": "^1.2.8",
-    "pulsar-client": "^1.12.0",
+    "pulsar-flex": "^1.1.1",
     "readline": "^1.3.0",
     "repl": "^0.1.3",
     "string-argv": "^0.3.2",
@@ -69,6 +70,7 @@
     "@babel/core": "^7.22.20",
     "@babel/plugin-proposal-unicode-property-regex": "^7.18.6",
     "@babel/preset-env": "^7.22.20",
+    "@jest/globals": "latest",
     "babel-loader": "^9.1.3",
     "copy-webpack-plugin": "^11.0.0",
     "jest": "^29.7.0",
