+++ src/index.ts
@@ -5,4 +5,5 @@ export * from './TemplateProcessor.js';
 export * as MetaInfoProducer from './MetaInfoProducer.js';
 export {default as JsonPointer} from './JsonPointer.js'
 export {stringifyTemplateJSON} from './utils/stringify.js'
+export {CliCoreBase} from './CliCoreBase.js';
